// apps/web/src/api/generated/models/quotationDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * APU Automotive Service Centre (APU-ASC) API
 * Automotive Service Management RESTful API Specifications
 * OpenAPI spec version: 1.0.0
 */
import type { QuotationLineDto } from './quotationLineDto'

export interface QuotationDto {
  id?: string
  quoteNumber?: string
  workOrderId?: string
  customerId?: string
  revision?: number
  status?: string
  notes?: string
  responseNotes?: string
  validUntil?: string
  subtotal?: number
  taxAmount?: number
  totalAmount?: number
  submittedAt?: string
  respondedAt?: string
  createdAt?: string
  updatedAt?: string
  items?: QuotationLineDto[]
}